emit ROM black box model
I think this is structured such that if no black boxing is selected,
the .bin file is generated as before; but if a black box is selected,
the system will emit an instance-specific ROM black box file that contains
a `case` statement that implements the ROM.

The `case` statement is less compatible with FPGA tools but at least
survives a compilation pass with synopsys (although the output is
non-optimal, it is at least correct -- the $readmemb implementation
is simply ignored by some synthesis flows). For more optimal implementation,
the black box verilog file can be swapped out for generated hard macros
that contain the ROM data when targeting an IC flow.