Patches to clean up warnings in synthesis
Some turds left over by codegen trigger warnings in synthesis. These
patches clear those out, but would break other forms of ROMs if they
were to exist in the code base.

There doesn't seem to be a good way to remove these warnings, because
of how the phases are separated. Charles would prefer to leave these
in because it means you don't destroy other behavior and they are
just warnings after all.

however, for the tape-out this project is targeting, the rule is
no warnings, and thus, I am monkey-patching out the warning-generating
code.